Implemented defaultIgnores
defaultIgnores defined as an array variable
setDefaultIgnores takses a `*IgnoresConfiguration` and populates it with
the contents of the defaultIgnores array
convertV7V8 to upgrade from v7 and add the default values

Default values are not populated on load as user may have chosen to
remove one or multiple values: this should not be reversed by
re-instationg. Defauls are set when upgrading or creating a new
configuration.